Update CPU.md
The preferred term is 'multi socket', as ever since the xeon scalable series, there is no longer any special distinction between dual socket and quad or octal socket.  From the OS's point of view, its all the same (you can even have odd numbers of sockets populated, like 3 or 5) and as long as you stay under macOS's thread limit, it will work all the same.  So I feel using multi socket is more appropriate, and this is the more modern terminology at any rate.

I also fixed an error regarding AppleMCEReporter.  MCE refers to machine check exception, which is used for (amongst other things) to report ECC memory errors.  Certain SMBIOSes, particularly the MacPro7,1 SMBIOS will cause it to be loaded, and if it loads on a system that doesn't have ECC memory, it will usually cause a panic.  However, it has no issues whatsoever as long as the system is equipped with ECC RAM.  I can personally attest to this across multiple versions of macOS and 2 generations of chipsets (C612 and C621) and CPUS (Broadwell-EP and Cascade Lake-EP Xeons).  macOS takes full advantage of ECC memory on these systems, and using APpleMCEReporterDisabler is not necessary on these systems, and will in fact compromise this fairly important functionality.